backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
httpx
redis
python-multipart
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
joblib==1.5.1
numpy==1.24.4
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
scikit-learn==1.7.1
scikit-surprise==1.1.4
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
